requirements.txt
#pip install opencv-python mediapipe pyserial pycaw
opencv-python
mediapipe
pyserial
pycaw